Backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
Django==3.0.6
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.9
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
webdriver-manager==2.5.3

